+++ src/types/ProcessError.ts
@@ -1,4 +1,5 @@
-import { isDevelopment, sanitizeMessage } from "../utils/standard.ts";
+import { isDevelopment } from "../utils/standard.ts";
+import { sanitizeMessage } from "../utils/internal.ts";
 
 export interface ProcessErrorPayload<TErrorDetails = unknown> {
 	code: string;